+++ modules/elementary_functions/tests/bench_rot90.m
@@ -7,12 +7,7 @@
 % SPDX-License-Identifier: LGPL-3.0-or-later
 % LICENCE_BLOCK_END
 %=============================================================================
-function varargout = ipermute(varargin)
-  narginchk(2, 2);
-  nargoutchk(0, 1);
-  A = varargin{1};
-  dimorder = varargin{2};
-  dimreordered(dimorder) = 1:numel(dimorder);
-  varargout{1} = permute(A, dimreordered);
-end
-%=============================================================================
+A = eye(3000);
+tic();
+B = rot90(A);
+toc()